feat: ExceptionHandler User

+++ src/main/java/com/example/ecommerce/services/UserService.java
@@ -4,39 +4,58 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.example.ecommerce.entities.User;
 import com.example.ecommerce.repositories.UserRepository;
-import com.example.ecommerce.services.exception.ResourceNotFoundException;
+import com.example.ecommerce.services.exceptions.DatabaseException;
+import com.example.ecommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> user = repository.findById(id);
 		return user.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User user) {
 		return repository.save(user);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
+		Optional<User> user = repository.findById(id);
+		if (user.isPresent()) {
+			try {
+				repository.deleteById(id);
+			} catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}
+		} else {
+			throw new ResourceNotFoundException(id);
+		}
 	}
-	
+
 	public User update(Long id, User data) {
-		User user = repository.getReferenceById(id);
-		updateUser(user, data);
-		return repository.save(user);
+		try {
+			User user = repository.getReferenceById(id);
+			updateUser(user, data);
+			return repository.save(user);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 
 	private void updateUser(User user, User data) {
